fix: rename clawd.invoke to openclaw.invoke + fix stdin merge
- Rename clawd_invoke.ts to openclaw_invoke.ts
- Change primary command name from 'clawd.invoke' to 'openclaw.invoke'
- Keep 'clawd.invoke' as deprecated alias for backward compatibility
- Update env vars: CLAWD_URL → OPENCLAW_URL, CLAWD_TOKEN → OPENCLAW_TOKEN (with fallbacks)
- Add --merge-stdin flag: merge first stdin JSON object fields into tool args (non-each mode)
- Add --spread flag: spread item fields into args instead of nesting (each mode)
- Update help text and error messages to reflect OpenClaw branding
- All existing tests pass, maintaining backward compatibility
